+++ src/super/i-block/modules/async-render/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-06-??)
+
+#### :bug: Bug Fix
+
+* Now all operations that are registered by asyncRender use only `async`, but not `$async`
+* Fixed a bug when rendered chunk are destroyed after creation when passed a custom async group
+
 ## v3.0.0-rc.204 (2021-06-23)
 
 #### :boom: Breaking Change